chore: Move submodule session out of module analysis into its own module.
