+++ hipcheck/src/main.rs
@@ -9,6 +9,7 @@ mod data;
 mod error;
 mod metric;
 mod report;
+mod session;
 mod shell;
 #[cfg(test)]
 mod test_util;
@@ -23,12 +24,12 @@ use crate::analysis::report_builder::PrReport;
 use crate::analysis::report_builder::Report;
 use crate::analysis::score::score_pr_results;
 use crate::analysis::score::score_results;
-use crate::analysis::session::Check;
-use crate::analysis::session::Session;
-use crate::analysis::session::TargetKind;
 use crate::context::Context as _;
 use crate::error::Error;
 use crate::error::Result;
+use crate::session::session::Check;
+use crate::session::session::Session;
+use crate::session::session::TargetKind;
 use crate::shell::Output;
 use crate::shell::Shell;
 use crate::shell::Verbosity;
