+++ hipcheck/src/session/session.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
-mod pm;
-mod spdx;
+// mod pm;
+// mod spdx;
 
 use crate::analysis::score::ScoringProviderStorage;
 use crate::analysis::AnalysisProviderStorage;
@@ -41,6 +41,8 @@ use crate::metric::MetricProviderStorage;
 use crate::report::Format;
 use crate::report::ReportParams;
 use crate::report::ReportParamsStorage;
+use crate::session::pm::detect_and_extract;
+use crate::session::spdx::extract_download_url;
 use crate::shell::Phase;
 use crate::shell::Shell;
 use crate::version::get_version;
@@ -329,7 +331,7 @@ fn resolve_source(
 
 			let command = &source_type.to_owned().kind;
 
-			let package_git_repo_url = pm::detect_and_extract(package, command.name().to_owned())
+			let package_git_repo_url = detect_and_extract(package, command.name().to_owned())
 				.context("Could not get git repo URL for package")?;
 
 			SourceRepo::resolve_repo(phase, home, package_git_repo_url.as_str()).map(|repo| {
@@ -344,7 +346,7 @@ fn resolve_source(
 			})
 		}
 		TargetKind::SpdxDocument => {
-			let download_url = spdx::extract_download_url(source)?;
+			let download_url = extract_download_url(source)?;
 			SourceRepo::resolve_repo(phase, home, &download_url).map(|repo| Source {
 				kind: SourceKind::Repo(repo),
 			})
