+++ hipcheck/src/session/spdx.rs
